+++ nautilus_core/model/src/enums.rs
@@ -344,13 +344,15 @@ pub unsafe extern "C" fn account_type_to_pystr(value: AccountType) -> *mut ffi::
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn account_type_from_pystr(ptr: *mut ffi::PyObject) -> AccountType {
-    AccountType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    AccountType::from_str(value)
+        .unwrap_or_else(|_| panic!("invalid `AccountType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -366,7 +368,7 @@ pub unsafe extern "C" fn aggregation_source_to_pystr(
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
@@ -374,8 +376,9 @@ pub unsafe extern "C" fn aggregation_source_to_pystr(
 pub unsafe extern "C" fn aggregation_source_from_pystr(
     ptr: *mut ffi::PyObject,
 ) -> AggregationSource {
+    let value = &pystr_to_string(ptr);
     AggregationSource::from_str(&pystr_to_string(ptr))
-        .expect("Error when parsing enum string value")
+        .unwrap_or_else(|_| panic!("invalid `AggregationSource` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -389,13 +392,15 @@ pub unsafe extern "C" fn aggressor_side_to_pystr(value: AggressorSide) -> *mut f
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn aggressor_side_from_pystr(ptr: *mut ffi::PyObject) -> AggressorSide {
-    AggressorSide::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    AggressorSide::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `AggressorSide` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -409,13 +414,15 @@ pub unsafe extern "C" fn asset_class_to_pystr(value: AssetClass) -> *mut ffi::Py
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn asset_class_from_pystr(ptr: *mut ffi::PyObject) -> AssetClass {
-    AssetClass::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    AssetClass::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `AssetClass` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -429,13 +436,15 @@ pub unsafe extern "C" fn asset_type_to_pystr(value: AssetType) -> *mut ffi::PyOb
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn asset_type_from_pystr(ptr: *mut ffi::PyObject) -> AssetType {
-    AssetType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    AssetType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `AssetType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -449,13 +458,15 @@ pub unsafe extern "C" fn bar_aggregation_to_pystr(value: BarAggregation) -> *mut
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn bar_aggregation_from_pystr(ptr: *mut ffi::PyObject) -> BarAggregation {
-    BarAggregation::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    BarAggregation::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `BarAggregation` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -469,13 +480,15 @@ pub unsafe extern "C" fn book_action_to_pystr(value: BookAction) -> *mut ffi::Py
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn book_action_from_pystr(ptr: *mut ffi::PyObject) -> BookAction {
-    BookAction::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    BookAction::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `BookAction` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -489,13 +502,15 @@ pub unsafe extern "C" fn book_type_to_pystr(value: BookType) -> *mut ffi::PyObje
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn book_type_from_pystr(ptr: *mut ffi::PyObject) -> BookType {
-    BookType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    BookType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `BookType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -509,13 +524,15 @@ pub unsafe extern "C" fn contingency_type_to_pystr(value: ContingencyType) -> *m
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn contingency_type_from_pystr(ptr: *mut ffi::PyObject) -> ContingencyType {
-    ContingencyType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    ContingencyType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `ContingencyType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -529,13 +546,15 @@ pub unsafe extern "C" fn currency_type_to_pystr(value: CurrencyType) -> *mut ffi
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn currency_type_from_pystr(ptr: *mut ffi::PyObject) -> CurrencyType {
-    CurrencyType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    CurrencyType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `CurrencyType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -549,7 +568,7 @@ pub unsafe extern "C" fn depth_type_to_pystr(value: DepthType) -> *mut ffi::PyOb
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
@@ -557,8 +576,10 @@ pub unsafe extern "C" fn depth_type_to_pystr(value: DepthType) -> *mut ffi::PyOb
 pub unsafe extern "C" fn instrument_close_type_from_pystr(
     ptr: *mut ffi::PyObject,
 ) -> InstrumentCloseType {
-    InstrumentCloseType::from_str(&pystr_to_string(ptr))
-        .expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    InstrumentCloseType::from_str(&pystr_to_string(ptr)).unwrap_or_else(|_| {
+        panic!("invalid `InstrumentCloseType` enum string value, was '{value}'")
+    })
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -574,13 +595,15 @@ pub unsafe extern "C" fn instrument_close_type_to_pystr(
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn depth_type_from_pystr(ptr: *mut ffi::PyObject) -> DepthType {
-    DepthType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    DepthType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `DepthType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -594,13 +617,15 @@ pub unsafe extern "C" fn liquidity_side_to_pystr(value: LiquiditySide) -> *mut f
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn liquidity_side_from_pystr(ptr: *mut ffi::PyObject) -> LiquiditySide {
-    LiquiditySide::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    LiquiditySide::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `LiquiditySide` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -614,13 +639,15 @@ pub unsafe extern "C" fn market_status_to_pystr(value: MarketStatus) -> *mut ffi
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn market_status_from_pystr(ptr: *mut ffi::PyObject) -> MarketStatus {
-    MarketStatus::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    MarketStatus::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `MarketStatus` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -634,13 +661,15 @@ pub unsafe extern "C" fn oms_type_to_pystr(value: OmsType) -> *mut ffi::PyObject
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn oms_type_from_pystr(ptr: *mut ffi::PyObject) -> OmsType {
-    OmsType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    OmsType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `OmsType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -654,13 +683,15 @@ pub unsafe extern "C" fn option_kind_to_pystr(value: OptionKind) -> *mut ffi::Py
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn option_kind_from_pystr(ptr: *mut ffi::PyObject) -> OptionKind {
-    OptionKind::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    OptionKind::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `OptionKind` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -674,13 +705,15 @@ pub unsafe extern "C" fn order_side_to_pystr(value: OrderSide) -> *mut ffi::PyOb
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn order_side_from_pystr(ptr: *mut ffi::PyObject) -> OrderSide {
-    OrderSide::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    OrderSide::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `OrderSide` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -694,13 +727,15 @@ pub unsafe extern "C" fn order_status_to_pystr(value: OrderStatus) -> *mut ffi::
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn order_status_from_pystr(ptr: *mut ffi::PyObject) -> OrderStatus {
-    OrderStatus::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    OrderStatus::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `OrderStatus` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -714,13 +749,15 @@ pub unsafe extern "C" fn order_type_to_pystr(value: OrderType) -> *mut ffi::PyOb
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn order_type_from_pystr(ptr: *mut ffi::PyObject) -> OrderType {
-    OrderType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    OrderType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `OrderType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -734,13 +771,15 @@ pub unsafe extern "C" fn position_side_to_pystr(value: PositionSide) -> *mut ffi
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn position_side_from_pystr(ptr: *mut ffi::PyObject) -> PositionSide {
-    PositionSide::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    PositionSide::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `PositionSide` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -754,13 +793,15 @@ pub unsafe extern "C" fn price_type_to_pystr(value: PriceType) -> *mut ffi::PyOb
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn price_type_from_pystr(ptr: *mut ffi::PyObject) -> PriceType {
-    PriceType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    PriceType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `PriceType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -774,13 +815,15 @@ pub unsafe extern "C" fn time_in_force_to_pystr(value: TimeInForce) -> *mut ffi:
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn time_in_force_from_pystr(ptr: *mut ffi::PyObject) -> TimeInForce {
-    TimeInForce::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    TimeInForce::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `TimeInForce` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -794,13 +837,15 @@ pub unsafe extern "C" fn trading_state_to_pystr(value: TradingState) -> *mut ffi
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn trading_state_from_pystr(ptr: *mut ffi::PyObject) -> TradingState {
-    TradingState::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    TradingState::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `TradingState` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -816,7 +861,7 @@ pub unsafe extern "C" fn trailing_offset_type_to_pystr(
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
@@ -824,8 +869,9 @@ pub unsafe extern "C" fn trailing_offset_type_to_pystr(
 pub unsafe extern "C" fn trailing_offset_type_from_pystr(
     ptr: *mut ffi::PyObject,
 ) -> TrailingOffsetType {
+    let value = &pystr_to_string(ptr);
     TrailingOffsetType::from_str(&pystr_to_string(ptr))
-        .expect("Error when parsing enum string value")
+        .unwrap_or_else(|_| panic!("invalid `TrailingOffsetType` enum string value, was '{value}'"))
 }
 
 /// Returns a pointer to a valid Python UTF-8 string.
@@ -839,11 +885,13 @@ pub unsafe extern "C" fn trigger_type_to_pystr(value: TriggerType) -> *mut ffi::
     string_to_pystr(&value.to_string())
 }
 
-/// Returns a pointer to a valid Python UTF-8 string.
+/// Returns an enum from a Python string.
 ///
 /// # Safety
 /// - Assumes `ptr` is borrowed from a valid Python UTF-8 `str`.
 #[no_mangle]
 pub unsafe extern "C" fn trigger_type_from_pystr(ptr: *mut ffi::PyObject) -> TriggerType {
-    TriggerType::from_str(&pystr_to_string(ptr)).expect("Error when parsing enum string value")
+    let value = &pystr_to_string(ptr);
+    TriggerType::from_str(&pystr_to_string(ptr))
+        .unwrap_or_else(|_| panic!("invalid `TriggerType` enum string value, was '{value}'"))
 }