+++ nautilus_trader/adapters/binance/spot/enums.py
@@ -37,6 +37,10 @@ class BinanceSpotPermissions(Enum):
     TRD_GRP_003 = "TRD_GRP_003"
     TRD_GRP_004 = "TRD_GRP_004"
     TRD_GRP_005 = "TRD_GRP_005"
+    TRD_GRP_006 = "TRD_GRP_006"
+    TRD_GRP_007 = "TRD_GRP_007"
+    TRD_GRP_008 = "TRD_GRP_008"
+    TRD_GRP_009 = "TRD_GRP_009"
 
 
 @unique