+++ tests/unit_tests/config/test_common.py
@@ -12,6 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
+
 import msgspec.json
 
 from nautilus_trader.config import ImportableConfig